Updated migration file for imie bigint

+++ db/migrate/20141102062308_add_imei_to_device.rb
@@ -1,5 +1,5 @@
 class AddImeiToDevice < ActiveRecord::Migration
   def change
-    add_column :devices, :imei, :integer
+    add_column :devices, :imei, :bigint
   end
 end
